fix: style and component changes 2
Product component is now ProductCard with different styles, Details component is now Product component
