+++ client/src/components/App/App.css
@@ -11,7 +11,7 @@ body {
 }
 .container {
   min-height: 100vh;
-  max-width: 1200px;
+  max-width: 1300px;
   margin: 0 auto;
   display: flex;
   align-items: center;
