+++ client/src/components/Home/Home.module.css
@@ -7,11 +7,9 @@
 .category-card {
   position: relative;
   flex: 1;
-  height: 500px;
-
+  height: 400px;
   display: flex;
   align-items: flex-end;
-
   background-size: cover;
 }
 .category-card span {
@@ -52,4 +50,16 @@
 
 .category-card:hover::before {
   background: rgba(0, 0, 0, 0.5);
+}
+.newest {
+  font-size: 1.8rem;
+  margin-block: 4rem;
+}
+.newest h3 {
+  margin-bottom: 1rem;
+}
+.newest-list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
 }
