--- client/src/components/Product/Product.css
@@ -1,95 +1,0 @@
-.product {
-  position: relative;
-  width: 300px;
-  padding: 20px;
-  background-color: #fff;
-  text-align: center;
-  height: 400px;
-  transition: transform 0.2s;
-  text-decoration: none; 
-  overflow: hidden;
-}
-
-.product-icons{
-  cursor: pointer;
-  font-size: 1.2rem;
-  color:#000;
-  opacity: 0;
-  transition: opacity 0.3s;
-}
-.product-icons span{
-  margin-left: 8px;
-  transition: color 0.3s;
-}
-
-.product-icons span:hover{
-  color:#049372;
-}
-
-
-
-.product:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 8px 12px rgba(0, 0, 0, 0.2);
-}
-
-.product img {
-  width: 200px;
-  height: 150px;
-}
-
-.product h3 {
-  font-size: 1.4rem;
-  margin: 10px 0;
-  color: #000000;
-  border-bottom: 2px solid #000000;
-}
-
-.product p {
-  color: #000000;
-  margin-bottom: 30px;
-}
-
-.price {
-  font-size: 1.3rem;
-  font-weight: bold;
-}
-
-.product-buttons {
-  position: absolute;
-  bottom: 10px;
-  left: 50%;
-  transform: translateX(-50%);
-  opacity: 0;
-  transition: opacity 0.3s;
-}
-
-.product:hover .product-buttons,.product:hover .product-icons {
-  opacity: 1;
-}
-
-.buy {
-  background-color: #049372;
-  color: #fff;
-  border: none;
-  padding: 8px 12px;
-  border-radius: 5px;
-  width:200px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  text-decoration: none;
-}
-
-.buy:hover {
-  background-color: #006442;
-}
-
-.description {
-  font-size: 1rem;
-}
-.product-icons {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  z-index: 1;
-}
