+++ client/src/components/Product/Product.module.css
@@ -1,0 +1,87 @@
+.product {
+  flex-grow: 1;
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+}
+.details-info {
+  display: flex;
+  flex-direction: row;
+}
+.furniture-buttons {
+  display: flex;
+  flex-direction: row;
+  gap: 4px;
+}
+.furniture-image {
+  flex: 1;
+}
+
+.furniture-image img {
+  max-width: 100%;
+  max-height: 300px;
+}
+
+.furniture-details {
+  flex: 1;
+  padding: 20px;
+}
+
+.furniture-details-page .buy-button {
+  background-color: #049372;
+  color: #fff;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 5px;
+  font-size: 18px;
+  cursor: pointer;
+  margin-top: 20px;
+  transition: background-color 0.3s;
+  display: flex;
+  align-items: center;
+}
+
+.furniture-details-page .buy-button:hover {
+  background-color: #006442;
+}
+
+.furniture-details-page .buy-button i {
+  margin-right: 5px;
+}
+
+.furniture-details-page h1 {
+  font-size: 28px;
+  margin-bottom: 10px;
+  color: #000000;
+}
+
+.furniture-details-page .price {
+  font-size: 24px;
+  font-weight: bold;
+  color: #000000;
+}
+
+.furniture-details-page .description {
+  margin-top: 10px;
+  font-size: 16px;
+}
+
+.furniture-details-page .add-to-cart-button,
+.furniture-details-page .add-review-button {
+  background-color: #049372;
+  color: #fff;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 5px;
+  font-size: 18px;
+  cursor: pointer;
+  margin-top: 20px;
+  transition: background-color 0.3s;
+}
+
+.furniture-details-page .add-to-cart-button:hover,
+.furniture-details-page .add-review-button:hover {
+  background-color: #006442;
+}
