+++ client/src/components/ProductCard/ProductCard.module.css
@@ -1,0 +1,30 @@
+.product-card {
+  position: relative;
+  width: 300px;
+  padding: 20px;
+  text-align: center;
+  height: 400px;
+  transition: all 300ms;
+  border: 2px solid rgb(201, 198, 198);
+}
+
+.product-card:hover {
+  border: 2px solid black;
+}
+
+.product-card img {
+  width: 200px;
+  height: 150px;
+  object-fit: contain;
+}
+
+.product-card h3 {
+  font-size: 1.2rem;
+  font-weight: normal;
+  margin: 10px 0;
+  color: #000000;
+}
+
+.product-card p {
+  font-size: 1.5rem;
+}
